Finished sort
This will pull all of the possible 3 word sentences that the
categories_xml file will give you.  It does not do it in the order of
that xml file, but that it because of the way the dictionary sorts
itself.  The code could be changed to go through the xml file a second
time to get the proper order.